ref: extract XML export logic to a new class
